1. deals and dealInfo for india.

+++ hasoffer.core/src/main/java/hasoffer/core/system/IAppService.java
@@ -33,7 +33,7 @@ public interface IAppService {
 
     OrderStatsAnalysisPO getOrderDetail(String orderId, String userId);
 
-    PageableResult getDeals(Long page, Long pageSize);
+    PageableResult getDeals(int page, int pageSize);
 
     PageableResult getDeals(Long page, Long pageSize, int temp);
